Fix preview on RN 76 (#712)
This PR fixes problem with preview functionality being broken on React
Native 76 (and on Fabric specifically).

The root cause was in the missing parameters we send to
AppRegistry.runApplication
On React Native 76, the renderer expects to receive `fabric` parameter
and also optionally `concurrentRoot` parameter that is placed under
`initialProps` objects. Because of the fact both of those parameters
were missing, rendering of the Preview container would throw weird error
about missing createElement and similar methods.

The fix is to capture `initialProps` and `fabric` attributes from the
original app wrapper, and forward them when calling
AppRegistry.runApplication later on.

### How Has This Been Tested: 
1. Open RN 76 example
2. Launch preview – before nothing would happen and there'd be a wall of
errors in the console. Now preview renders ok. Note that there's another
preview related issue, so you need to make sure `preview` call and the
component it renders are defined on the same line.