Properly match line number to fix multi-line previews (#713)
This PR fixes an issue with previews that span multiple lines, for
example:
```tsx
preview(
  <Button
    title="Button"
    onPress={() => {
      console.log('Hello');
    }}
  />,
);
```

Opening this issue would previously result in no interaction and an
error about missing name would be displayed in the console.

The root cause of the issue is the mismatch between line numbers which
codelens provider uses for requesting preview render, and the line
number the preview uses to register. In the previous implementation when
rendering preview, we'd request it using filename and line number of
where `preview` instruction was used. On the other hand, when
registering preview components, we use the line where component is
defined (this is a line number extracted by the JSX transformer)

This worked ok for components that would be placed in the same line, for
example:
```tsx
preview(<NiceButton color="blue" />);
```

But, for multi-line previews as shown above, the tag can be in a
different line.

The fix is to update codelens provider to not just match `preview(` but
also look for the opening JSX tag ('<' character). We then use the
location of the opening JSX tag as the line where the preview is
expected.

Fixes  #690

### How Has This Been Tested: 
1. Open RN 76 example where multi-line preview is used on MainScreen
2. Launch the multi line preview
3. Add another single-line one, and make sure it works
4. Add preview that's commented out to verify it's not being picked up
by the codelens provider