Rerender preview if its function changes (#726)
This PR fixes an issue when changes made to `preview` component directly
wouldn't trigger UI updates.

The reason why this wasn't happening is due to the limitation of
fast-refresh, which, when happens, replaces the implementation of the
previous component. As a consequence, the root Preview component doesn't
see the previewed one got updated.

In this PR we're adding `renderTrigger` callback that the root Preview
component uses to register for updates in the preview map. When
registering the preview, we check if a one already exists, and trigger
update for it.

Caveat: note that with this approach the root component that is used by
the preview won't persist state. I don't see how can we workaround this,
but it doesn't seem too limiting at this point.

Fixes #725

### How Has This Been Tested: 
1. Run expo-router example, use preview from UglyButton.tsx
2. Change color prop passed to UglyButton directly – before this change
the component wouldn't update.