Add exception handler for source-map originalPositionFor calls (#773)
This PR adds try/catch block around `originalPositionFor` calls. From
telemetry we see this method throwing unhandled exceptions. While we
don't yet understand the reason for those errors, we can always fallback
to using non-translated lines. This will still degrade the experience
when this error happen, but at least wouldn't prevent logs from
appearing in the console tab, will offload our telemetry as we're likely
missing some important signals because of this exception, and may
hopefully allow us to troubleshoot the root cause of this issue.

### How Has This Been Tested: 
This hasn't been tested as we're not able to reproduce the problem with
invalid source map