Make fingerprint run less often (#794)
This PR updates fingerprint throttling such that it runs less often. In
practice it doesn't seem like the delay of several seconds makes for an
important difference when detecting native changes. When updating native
changes most of the time, the refresh would trigger exception regardless
of how soon we're able to detect them based on the fingerprint. What is
important is to give a clear indication that this was due to native
changes getting updated.

On the other hand, to frequent fingerprint calls are annoying in cases a
lot of files are invalidated at the same moment (for example when
installing node_modules) in which case this process incurs a significant
CPU load.

### How Has This Been Tested: 
1) run one of the test apps
2) add some comments to Podfile
3) see the native changes dialog pop up